Add graphql types to air-discount project as well. Fix some Float -> Int in types.

+++ apps/web/graphql/schema.ts
@@ -131,7 +131,7 @@ export type ArticleGroup = {
 export type ArticleSubgroup = {
   __typename?: 'ArticleSubgroup'
   title: Scalars['String']
-  importance?: Maybe<Scalars['Float']>
+  importance?: Maybe<Scalars['Int']>
   slug: Scalars['String']
 }
 
@@ -308,7 +308,7 @@ export type Article = {
   shortTitle: Scalars['String']
   intro: Scalars['String']
   containsApplicationForm?: Maybe<Scalars['Boolean']>
-  importance?: Maybe<Scalars['Float']>
+  importance?: Maybe<Scalars['Int']>
   body: Array<Slice>
   category?: Maybe<ArticleCategory>
   group?: Maybe<ArticleGroup>
@@ -537,7 +537,7 @@ export type ContentItem = {
   categorySlug?: Maybe<Scalars['String']>
   categoryDescription?: Maybe<Scalars['String']>
   containsApplicationForm?: Maybe<Scalars['Boolean']>
-  importance?: Maybe<Scalars['Float']>
+  importance?: Maybe<Scalars['Int']>
   group?: Maybe<Scalars['String']>
   subgroup?: Maybe<Scalars['String']>
   groupSlug?: Maybe<Scalars['String']>
